added moved retrofit instance to app class
