+++ app/src/main/java/com/kodeco/android/countryinfo/CountryApp.kt
@@ -1,0 +1,30 @@
+package com.kodeco.android.countryinfo
+
+import android.app.Application
+import com.kodeco.android.countryinfo.network.CountryService
+import com.kodeco.android.countryinfo.repositories.CountryRepository
+import com.kodeco.android.countryinfo.repositories.CountryRepositoryImpl
+import com.squareup.moshi.Moshi
+import com.squareup.moshi.kotlin.reflect.KotlinJsonAdapterFactory
+import retrofit2.Retrofit
+import retrofit2.converter.moshi.MoshiConverterFactory
+
+class CountryApp: Application() {
+
+    lateinit var repository: CountryRepository
+
+    override fun onCreate() {
+        super.onCreate()
+        val moshi = Moshi.Builder()
+            .add(KotlinJsonAdapterFactory())
+            .build()
+
+        val retrofit = Retrofit.Builder()
+            .baseUrl("https://restcountries.com/v3.1/")
+            .addConverterFactory(MoshiConverterFactory.create(moshi))
+            .build()
+
+        val service: CountryService = retrofit.create(CountryService::class.java)
+        repository = CountryRepositoryImpl(service)
+    }
+}
