+++ app/src/main/java/com/kodeco/android/countryinfo/models/CountryFlags.kt
@@ -1,6 +1,9 @@
 package com.kodeco.android.countryinfo.models
 
+import android.os.Parcelable
 import com.squareup.moshi.JsonClass
+import kotlinx.parcelize.Parcelize
 
+@Parcelize
 @JsonClass(generateAdapter = true)
-data class CountryFlags(val png: String, val svg: String)
+data class CountryFlags(val png: String, val svg: String): Parcelable
