+++ app/src/main/java/com/kodeco/android/countryinfo/models/CountryName.kt
@@ -1,6 +1,9 @@
 package com.kodeco.android.countryinfo.models
 
+import android.os.Parcelable
 import com.squareup.moshi.JsonClass
+import kotlinx.parcelize.Parcelize
 
+@Parcelize
 @JsonClass(generateAdapter = true)
-data class CountryName(val common: String)
+data class CountryName(val common: String): Parcelable
