+++ app/src/main/java/com/kodeco/android/countryinfo/ui/components/CountryInfoList.kt
@@ -19,6 +19,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -38,7 +39,7 @@ fun CountryInfoList(
     onCountryRowTap: (Int) -> Unit,
     onTapAbout: () -> Unit
 ) {
-    var selectedCountry: Country? by remember { mutableStateOf(null) }
+    var selectedCountry: Country? by rememberSaveable { mutableStateOf(null) }
 
     Column {
         Row(
